src/main/java/org/stellar/sdk/scval/ScvAddress.java
package org.stellar.sdk.scval;

import lombok.EqualsAndHashCode;
import lombok.RequiredArgsConstructor;
import lombok.Value;
import org.stellar.sdk.Address;
import org.stellar.sdk.xdr.SCVal;
import org.stellar.sdk.xdr.SCValType;

/** Represents an {@link SCVal} with the type of {@link SCValType#SCV_ADDRESS}. */
@Value
@RequiredArgsConstructor
@EqualsAndHashCode(callSuper = false)
class ScvAddress extends Scv {
  private static final SCValType TYPE = SCValType.SCV_ADDRESS;

  Address value;

  @Override
  public SCVal toSCVal() {
    return value.toSCVal();
  }

  public static ScvAddress fromSCVal(SCVal scVal) {
    if (scVal.getDiscriminant() != TYPE) {
      throw new IllegalArgumentException(
          String.format(
              "invalid scVal type, expected %s, but got %s", TYPE, scVal.getDiscriminant()));
    }

    return new ScvAddress(Address.fromSCVal(scVal));
  }
}
